requirements.txt
streamlit
pandas
numpy
altair
snowflake-snowpark-python
snowflake-connector-python>=3.0.0
snowflake-cli-labs
fredapi
boto3
numpy
pytest
python-dotenv
jinja2

